Add standard repo files, CI workflows
We ignore sln files on the root dir since we use a common standard
CI script that just runs `dotnet format` on the root, so there must
be only a single .sln or .csproj. Since we don't have more than one
project, it should be fine to just ignore solution files.