:cat: Add IAsyncEnumerable<T> extension methods
The native way of consuming `IAsyncEnumerable<T>` is obviously an `await foreach` loop. Sometimes, however (especially in tests, or when dealing with few records) it's useful to be able to quickly materialize the enumerable (asynchronously) to an array, a dictionary, a set or a list.

This introduces the async versions of the common enumerable extension methods such as `ToArrayAsync`, `ToDictionaryAsync`, `ToHashSetAsync` and `ToListAsync`.