🐱 Add IServiceCollection.AddLazy() extension method
This allows resolving any service lazily using `Lazy<T>` as a dependency instead of the direct service type.

```csharp
services.AddLazy();
services.AddSingleton<IMyExpensiveService>(() => // expensive creation here);
```

A component could declare a dependency like:

```csharp
public MyController(Lazy<IMyExpensiveService> service)
```

And the registration side of the service wouldn't change at all.

The benefit of depending on the Lazy<T> is that the controller (in this example) might not end up needing the expensive service at all, for example. And using the lazy in this case, would avoid entirely the service creation.